Add Basic Auth for multiple users
As we are using paid APIs, a fully open instance could lead to unchecked spendings on the account of the instance's owner.

With basic auth, one can deploy their public instance and open it to their friends, or set up an instance for a workshop. Or they can keep their instance private just for them.

This a very basic user management scheme, based on manually  set environment variables for each user.