Scroll to reordered block in auto miniview
When entering the miniview automatically, on finishing the sort, the view will scroll to center the reordered group or groups.
It gives a much better experience and feels like what the user expect.
If entering the miniview via the button, the view still scroll back to the position we were when entering. If we have sorted blocks, then the position might appears different of course.

Add a small delay to the scroll to ensure the transition is finished. Without the delay the scroll to element wasn’t working consistently. (the scroll to position seemed to work, but the added delay isn't sensible and may avoid some edge cases)