We can import images in a dedicated block
This new block display the image and store it as a js fil image in its result. This will allows other blocks to use this result for: image-to-image, controlNet or image-to-text ('vision model').
The use can either add a block and drop/import an image on it, or directly drop one or several images in the container zone, which will create the blocks automatically.