Versions using kvdex history; list user's apps
Use the kvdex history. This makes it simpler to write and retrieve the document. History is stored in a parallel index for when needed.
That's also makes it trivial again to retrieve the list of apps created by an user (when previously all versions of the app would have been listed as an app)
Also rename kv-storage to apps as the module that handles apps storage, history, retrieval and results.