Remove the user facing 'image tags'
these tags are quick (temporary) way to give feedback to users that their image is indeed referenced and to make the changes taken into account.
We remove them from the prompt to avoid unintended effects. In particular it seems to throw GPT4 vision off, making it even believe that no image is attached. It can also influence image generation via the name of the group itself.