Add capture thumbnail utility
add captureThumbnail function that uses HTML2Canvas to render the visible part of the app as a thumbnail, respecting the scroll position.
As the meshbackground buffer is cleared after rendering, and thus appears blanks when we HTML2Canvas try to render or when we manually try copy it, we add a function that render afresh and returns an blobUrl copy of the background. We then swap the background for the static image in the onclone callback, applying the same absolute positioning.
if window.DEBUG is true, the capture will be saved on disk