Cient persists and loads imported image
Why? Imported image can have a structural usage in an app, for example a controlnet guiding image to draw a face.

When an image is dropped/added to a imported image block, the image is sent to the server to be persisted.
The client gets a hash digest back for the image, that is then added to the group data, and the groups structure is persisted to save the reference for future loading of the app
On loading an app, imported image groups are checked for image hash values, and we call the server to get the images back, set it in the document and in the group's result property for dependent groups to use

This commit also replace the foreach that created all the group elements on loading with a parallel map returning promises. The changes was needed because of the async nature of the call to fetch the imported images back from the server. We wait for all promises to resolves before calling to update all the groups.