Data change in temp entry: mark structure unchanged
Marking the app structure as changed if data text changed in a group not set as a temporary data entry block.
When user input new data in an entry block, the corresponding group is considered unchanged in terms of structure, and that will not trigger a new persisting of the groups structure on server.
This will also avoid a remix when using another user app throught entry blocks