Throttle groups persist
Saving a new change to groups only after a cooldown of 30 seconds or 10 unsaved changes, whatever comes first.
This should cut the number of versions by 70%.
To balance out the relatively long cooldown, we persist when the tab is hidden (especially important on mobile) and when the tab is closed.