improve list handling: keep position over regen
We now keep the picked position in a list if possible. This is useful for users when using list format to split a list items through several blocks. Ex: list generated => reference the full list and pick item 1 in a static text block, item 2 into another, etc.

We delay the extract the call to update the  groups referencing the group with new results from the sendRequestsForGroup function, and bring it back in handleInputChange. This allow us to either update instantly or format first