Tooltips
Use tippy.js to add toggable tooltips on tool button icons. Tooltips are off by default.
Also add permanent Aria Descriptions for buttons, in addition to the Aria Label
We now emit an event when a group element is added or removed, and tooltips are recreated when the group elements change to match the existing elements.