Fallback on LLAVA 1.6 if GPT4-vision fail
GPT4-vision is strongly rate limited preview.

We fallback on LLAVA 7B and 32B on Replicate if the call to GPT4-vision fail (for example because of the rate limiting)

LLAVA 7B Mistral seems quite effective for a range of description tasks. 32B is activated using quality mode, and is quite slower