Fix remixing when deleting other user's app
Deleting an app is actually persisting an empty version of the app, using the app id.
As trying to persist an app from an other user (using their app id) will always create a remix, by trying to delete others' apps, we just add empty remix to our own apps.
This fix the issue server side, by never allowing to try and persist and empty 'remix', but it should be handled on the client side too, by hiding the delete UI when the current user is not the user who created the apps displayed