recover app endpoint
Recover an app by founding its last non-empty version and storing this version as the newest version of the app. Recovering a non-empty app will just store the current version again.

Instead of recovering by persisting from the client, the recover endpoint allows to recover an app by just sending the app id.

Only the creator of an app can recover this app.