imported image: temporary entry setting
Add the open / temporary entry entry setting buttons to imported image blocks.
The buttons are added to the current clear buttons. All buttons are always showed, for a simple and less surprising behavior.
When entry is set to temporary entry, the image hash is not created and saved in the groups and the groups are not persisted on adding an image to block.