Clone button replace add block buttons when loading another user's app
When loading another's user app, the add block buttons are hidden and replaced with a [clone and remix app] button.

Clicking on the button will persist the current app, which will create a copy of the app for the current user.

The persist groups function will detect it's an new app originally from a different user and reveal the add block buttons (in addition to the previous remix alert)