Reorder: groups have explicit index value
instead of recreating the full map after reordering the groups, we just rewrite the index property of each group.
Index properties are initially set on
- loading groups: using the groups array order
- creating a new group: as new groups are added at the end, using the size of the groups map

Index properties are not persisted, instead the groups map is converted into an array, sorted by index value: groups array order is the order to persist and use on load. It makes for a simpler persistence format.