Interaction State: Move lock state setting out of loading to display
As the lock state is now automatically set when using other's apps, and not when using our own apps, we don't set this state as a value in the group data, but we set it as a pure UI status when displaying the interaction state of group.
What decides the UI lock state is the ownership of the app, rather than an explicitly locked state.