Reset the interaction state of cloned apps
Cloning an app is just persisting the groups under the user own name.
After persisting, we check if the app that we have persisted is a clone (new and previously by a different creator), and if it's a clone we refresh the display of all groups' interaction states: they will pass from locked to open for the open groups, and stay on entry for the entry groups.