Replace blobURIs with a custom esquisse result reference locator
blob URIs are client only and transient, so there's no point in saving them in the result
We identify of the blob URIs in the text results, and replace them with a custom scheme:
esquisse:result/[reference]
where reference is the name of the image group result referenced
When displaying the results, we can use this special locator to get the data URI for the image and use it.
It's also less prone to surprises than adding a data attribute like before, because if the URI is removed by a text transform, the user can immediatly see that the image is not shown: what is shown in result and app can match again.