Use itemized datatext and small iframes or img tags to preview references
References are rendered as a series of either small iframes or img tags in th referenced results preview. Text entered by the user that is not a reference is added as spans.
References that are invalid or not ready are added as spans with the content set to their name and with the appropriate style (red for invalid, orange for not ready).

As instruct LLMs tend to return markdown formatted answers, we first render the results from static or text gen groups with marked, then sanitize and ultimately escape <>