Fuse data and transform into just data
Users are confused by the two inputs, and with the new reference preview mechanism, it has lost its importance as a way to delineate data and instructions.

pros:
-we can now use a bigger single input
- no more not functioning references in the transform: users added them naturally and didn't even noticed they didn't work.

Issue: the static and text generation blocks look very similar visually.

Retro compatibility:
When loading an existing app, any transform text is appended to the data text.

The transform is always packaged as empty, and will be entirely removed in later versions.

We update the Esquisse version to distinguish the saved apps, as this is a big change.