UX: fix user can try to select in referenced-result-text to no avail
Handling Mousedown gives us an immediate feedback of focusing data input when starting a click anywhere in the data input element, except inside a rendered result iframe.

It's useful because it avoids the mentally surprising experience of instinctively selecting a text in the rendered referenced results and loosing the selection when the pointer button is released.

Click, in turn, focus the data input textArea when clicking anywhere, including a rendered result iframe, giving us an acceptable fallback for these cases.