Sorted groups display their formatted results anew
fix result iframe content being dropped on sorting.

reordering.js emit a custom sorted event on each group that has ended being sorted

groups listen to this event and display their formatted results when receiving it.

We also make the displayFormattedResults function iso on its behavior with list: calling it on a group that is already in list mode now work as expected and display the exact same list and position. Previously it was trying to parse the selected item stored as result as a list, and failed.

This also fix an issue when calling it on an image group, where it cleared the group.result (it now just do nothing, which is the correct behavior)