Store app file versions on S3
Storing whole versions with all the groups on KV is wasting a lot of operations and space, in what amount to storing app read only files, especially as we store the history without any optimizations except some client-side cooldown between saves.

We now store versions full content on S3, only keeping the metadata needed for retrieval (like for results), including a versionhash that is computed by storeApp before sending the app version to be uploaded on S3.

retrieveAppVersionFromSource and retrieveAppFromSource handle the new path and the legacy path:
- they are passed the data retrieved from KV,
-  depending on the presence of versionhash, they fetch the full app version from S3 or use the KV data as is.