Remove IMG URL from prompts and filenames
We don’t want IMG tags from the combinedresult added to prompts sent via sendRequestsForGroup and filenames of generated images downloaded by the user. It pollutes the prompt sent, and makes the filename unhelpfu.

In the current code, the prompt and filename is based on the group.result, but group.result is built as an intermediate result that is to be used by other groups, and it’s inadequate to use it for other purposes.

Instead we use the itemizedDataText list built by getReferencedResultsAndCombinedDataWithResults instead, and build the right prompt text and the right filenames:

- build the text prompt that is sent via sendRequestsForGroup using the itemized data text and the following rules:
skip the image references entirely (item of type ‘IMAGE’)
only add the item of contenType ‘HTML’ if isReady == true and isValid == true
add the items of contentType ‘TEXT’.

- build the filename of downloaded images using the itemized data text. Add, after the shortened prompt: either [image to image] or [controlnet], if an image has been used for the generation process.