Image gen fallback models; fix types
Add a fallback handling similar to the fallback in gpt.js module

Models in MODELS can have a fallback property.

Add fallback to fal for segmind controlnet models as segmind tends to have downtimes as it grows. Fal endpoints are quite slower on this model, tho.

Also tidy the types and the image result object that is passed up