Mesh background render the full view
previously the mesh background canvas was sized to fit the groups, but with apps with only a couple of groups, that lead to the canvas not covering the full view, and the background being visible.
It was a particular issue for programmatic captures of results thumbnails
Now the meash background canvas is at least the height of the app groups zone and as high as the scroll height of the groups's container.