Display the generated title and summary to user with sharing link
In the 'result ready to share' notification, show title and summary generated on the server, and the locally generated thumbnail: user can see the generated title to help them find their result later, and they get an overview of what they are sharing, as it is also the open graph content.

On successfully storing the result, the server is now returning not only the result id, but also the generated title and generated summary.